Removed header from HomePage, renamed files, removed clickOn() method from BasePage, removed unnecessary waitForLoadState() methods
