+++ src/po/pages/product.page.js
@@ -1,4 +1,4 @@
-import BasePage from "./BasePage";
+import BasePage from "./base.page";
 
 class ProductPage extends BasePage {
   constructor(page) {
