+++ src/tests/product-details.spec.js
@@ -1,13 +1,13 @@
 import { test, expect } from "@playwright/test";
 import { assert } from "chai";
 
-import RegisterPage from "../po/pages/RegisterPage";
-import LoginPage from "../po/pages/LoginPage";
-import FavoritesPage from "../po/pages/FavoritesPage";
-import HomePage from "../po/pages/HomePage";
-import ProductPage from "../po/pages/ProductPage";
+import RegisterPage from "../po/pages/register.page";
+import LoginPage from "../po/pages/login.page";
+import FavoritesPage from "../po/pages/favorites.page";
+import HomePage from "../po/pages/home.page";
+import ProductPage from "../po/pages/product.page";
 
-import { testUser } from "../data/userData";
+import { testUser } from "../data/user.data";
 
 test.describe("product details page", () => {
   test.beforeEach(async ({ page, context, baseURL }) => {
@@ -23,7 +23,7 @@ test.describe("product details page", () => {
     });
 
     try {
-      await registerPage.clickOn(registerPage.registerBtn);
+      await registerPage.registerBtn.click();
       await expect(page).toHaveURL(baseURL + "/auth/login");
       console.log("Account created");
     } catch (err) {
@@ -33,7 +33,7 @@ test.describe("product details page", () => {
     // Given the user is logged
     await loginPage.navigateTo(baseURL + "/auth/login");
     await loginPage.enterCredentials("email@example.com", "123_Tests");
-    await loginPage.clickOn(loginPage.loginSubmit);
+    await loginPage.loginSubmit.click();
     await loginPage.waitForURL(baseURL + "/account");
     const accountURL = page.url();
     assert.equal(accountURL, "https://practicesoftwaretesting.com/account");
@@ -57,8 +57,7 @@ test.describe("product details page", () => {
     await homePage.clickOnProduct("Combination Pliers");
 
     // When the User adds the Combination Pliers to Favorites
-    // await productPage.addToFavoritesClick();
-    await productPage.clickOn(productPage.addToFavorites);
+    await productPage.addToFavorites.click();
 
     // Then the message "Product added to your favorites list." is displayed
     const addedToFavoritesVisible = await productPage.alertIsVisible("Product added to your");
@@ -74,7 +73,7 @@ test.describe("product details page", () => {
     assert.isTrue(combinationPliersIsFavorite, "Expected Combination Pliers to be visible.");
 
     // (Remove the product for future re-test)
-    await favoritesPage.clickOn(favoritesPage.deleteBtn);
+    await favoritesPage.deleteBtn.click();
     const noFavsisVisible = await favoritesPage.isVisible(favoritesPage.noFavoritesMessage);
     assert.isTrue(noFavsisVisible, "Expected 'There are no favorites yet.' message to be visible");
   });
@@ -92,7 +91,6 @@ test.describe("product details page", () => {
     await homePage.clickOnProduct("Long Nose Pliers");
 
     // When the Long Nose Pliers page has the "Out of stock" message
-    // const outOfStockIsVisible = await productPage.outOfStockIsVisible();
     const outOfStockIsVisible = await productPage.isVisible(productPage.outOfStock);
     assert.isTrue(outOfStockIsVisible, "Expected Out Of Stock message to be visible.");
